Rename company DTO and company category DTO
Refactor company dto and company category dto as they are used in Response dtos repetitively
