+++ src/main/java/com/blind/company/api/dto/request/CreateCompanyRequest.java
@@ -1,0 +1,11 @@
+package com.blind.company.api.dto.request;
+
+import com.blind.company.api.dto.shared.CompanyCategoryDto;
+
+import java.util.UUID;
+
+public class CreateCompanyRequest {
+    private UUID id;
+    private String name;
+    private CompanyCategoryDto companyCategory;
+}
