+++ src/main/java/com/blind/company/api/dto/response/CreateCompanyResponse.java
@@ -1,0 +1,5 @@
+package com.blind.company.api.dto.response;
+
+public class CreateCompanyResponse {
+    private String Id;
+}
