+++ src/main/java/com/blind/company/api/dto/shared/CompanyCategoryDto.java
@@ -1,4 +1,4 @@
-package com.blind.company.api.dto.response;
+package com.blind.company.api.dto.shared;
 
 import com.blind.shared.api.BaseDtoResponse;
 import lombok.*;
@@ -9,8 +9,7 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class CompanyResponse extends BaseDtoResponse {
+public class CompanyCategoryDto extends BaseDtoResponse {
     private UUID id;
     private String name;
-    private CompanyCategoryResponse companyCategory;
 }
