+++ src/main/java/com/blind/company/api/mapper/CompanyMapper.java
@@ -1,7 +1,8 @@
 package com.blind.company.api.mapper;
 
-import com.blind.company.api.dto.response.CompanyCategoryResponse;
-import com.blind.company.api.dto.response.CompanyResponse;
+import com.blind.company.api.dto.request.CreateCompanyRequest;
+import com.blind.company.api.dto.shared.CompanyCategoryDto;
+import com.blind.company.api.dto.shared.CompanyDto;
 import com.blind.company.domain.Company;
 import com.blind.company.domain.CompanyCategory;
 import org.mapstruct.Mapper;
@@ -14,7 +15,9 @@ import java.util.List;
 public interface CompanyMapper {
     CompanyMapper INSTANCE = Mappers.getMapper(CompanyMapper.class);
 
-    CompanyResponse companyToDto (Company company);
-    CompanyCategoryResponse companyCategoryToDto (CompanyCategory companyCategory);
-    List<CompanyResponse> companyListToDto(List<Company> companyList);
+    CompanyDto companyToDto(Company company);
+    Company dtoToCompany(CreateCompanyRequest companyDto);
+    CompanyCategoryDto companyCategoryToDto(CompanyCategory companyCategory);
+    CompanyCategory dtoToCompanyCategory(CompanyCategoryDto companyCategoryDto);
+    List<CompanyDto> companyListToDto(List<Company> companyList);
 }
