+++ src/test/java/com/blind/api/mapping/MappingTests.java
@@ -1,6 +1,6 @@
 package com.blind.api.mapping;
 
-import com.blind.company.api.dto.response.CompanyResponse;
+import com.blind.company.api.dto.shared.CompanyDto;
 import com.blind.company.api.mapper.CompanyMapper;
 import com.blind.company.domain.Company;
 import com.blind.company.domain.CompanyCategory;
@@ -25,7 +25,7 @@ public class MappingTests {
         CompanyCategory companyCategory = new CompanyCategory(UUID.randomUUID(), "CompanyCategory");
         Company company = new Company(UUID.randomUUID(), "Company", companyCategory);
 
-        CompanyResponse companyDto = companyMapper.INSTANCE.companyToDto(company);
+        CompanyDto companyDto = companyMapper.INSTANCE.companyToDto(company);
 
         assertNotNull(companyDto);
         assertEquals(companyDto.getName(), company.getName());
